feat: initial settings and refactor
