+++ assets/components/src/with-wizard/style.scss
@@ -7,6 +7,7 @@
 
 // Reset Padding of the Admin Page
 
+.toplevel_page_newspack-dashboard:not( .menu-top ),
 .toplevel_page_newspack:not( .menu-top ),
 body[class*='newspack_page_newspack-'],
 body[class*='admin_page_newspack-'] {
