+++ assets/wizards/newspack/components/icons/dashboard.tsx
