+++ assets/wizards/newspack/components/icons/gift.tsx
