+++ assets/wizards/newspack/components/icons/mail.tsx
